Refactor scanner context location and imports
