+++ frontend/src/app/providers/providers.tsx
@@ -6,7 +6,7 @@ import { SidebarProvider } from "@/app/providers/sidebar-provider";
 import { ToasterProvider } from "@/app/providers/toaster-provider";
 import { UserContextProvider } from "@/app/providers/user-provider";
 import { ReactScan } from "@/app/providers/react-scan";
-import { ScannerProvider } from "@/components/custom/scanner-context";
+import { ScannerProvider } from "@/context/scanner-context";
 
 export function Providers({ children }: { children: ReactNode }) {
   return (
